refactor: change the file name of isValidValue
