+++ src/app/(auth)/sign-up/_components/sign-up-form/use-sign-up-form.tsx
@@ -7,7 +7,7 @@ import { useModal } from '@/components/modal/use-modal'
 import { signUpSchema } from '@/schemas/request/auth'
 import { createFormErrorsSchema } from '@/utils/form/create-form-errors-schema'
 import { useHandleFormErrors } from '@/utils/form/use-handle-form-error'
-import { isValidValue } from '@/utils/type-guard/is-valid-data'
+import { isValidValue } from '@/utils/type-guard/is-valid-value'
 import { signUp } from './sign-up.api'
 import type { ErrorObject } from '@/types/error'
 import type { HttpError } from '@/utils/error/custom/http-error'
