+++ src/components/pages/account-page/current-user-info/editable-avatar/avatar-editor/use-avatar-editor.tsx
@@ -9,7 +9,7 @@ import { HttpError } from '@/utils/error/custom/http-error'
 import { createFormErrorsSchema } from '@/utils/form/create-form-errors-schema'
 import { useHandleFormErrors } from '@/utils/form/use-handle-form-error'
 import { useRedirectLoginPath } from '@/utils/login-path/use-redirect-login-path'
-import { isValidValue } from '@/utils/type-guard/is-valid-data'
+import { isValidValue } from '@/utils/type-guard/is-valid-value'
 import { changeAvatar } from './change-avatar.api'
 import { changeAvatarSchema } from './change-avatar.schema'
 import type { SubmitHandler } from 'react-hook-form'
