+++ src/utils/type-guard/is-valid-value.ts
@@ -2,7 +2,7 @@ import type z from 'zod'
 
 export function isValidValue<T extends z.ZodTypeAny>(
   schema: T,
-  value: unknown
+  value: unknown,
 ): value is z.infer<T> {
   return schema.safeParse(value).success
 }
